requirements.txt
redis
crontab
celery
environ
pytz
chardet
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
dj-database-url
click==8.1.7
django-redis
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
idna==3.7
kombu==5.4.0
packaging==24.1
pillow==10.4.0
pandas==2.2.3
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
apscheduler
fastapi
requests
pandas 
openpyxl